04-financial-report-rag/requirements.txt
anthropic>=0.40.0
streamlit>=1.35.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pdfplumber>=0.11.0
numpy>=1.26.0
ragas>=0.1.0
datasets>=2.20.0
python-dotenv>=1.0.0
